: Animation End Event and TextBox ancestor
I did two things related to "listeners". First, I made an
AnimationListener. Each animation can have up to one AnimationListener.
When the Animation either loops or comes to an end, it will notify its
listener. This can be very useful if some event must happen after the
animation finishes playing. A recommended use is to make the MenuManager
implementation an AnimationListener so it can react when necessary.

The other minor change I made was making text input more generic.
Previously, the menu manager would only pass key events onto subclasses
of TextBox through the append and delete methods. This would obviously
require an future text variant components to subclass TextBox, which may
not be desirable especially considered all the other methods and
features that TextBox has that may not be applicable for the future
component. Thus, there is a TextInputComponent interface that is a
superclass for HotkeyTextComponent. The menu manager will still handle
the key typing events as before, but instead of passing exclusively to
the append and remove methods of TextBox, the input is passed to the
appendMessage and removeMessage methods of TextInputComponent, which
TextBox then overrides.