Partition panels, text box ancestor, animation listening
There were a lot of small improvements along with a few new features and
some general polishing. Partition panels give the ability to change at
run-time how much space a component will take up on the screen. This is
particularly useful if some components need to be invisible sometimes,
but the partition panel is cleaner than adding and removing said
component. There can be a partition between a scroll bar and its
component, and the scroll bar can be hidden when it is uneditable.

A new TextBox ancestor was made. Now components do not have to inherit
from TextBox to receive key input events. They just have to implement
TextInputComponent.
Similarly with ScrollBars and ScrollableComponent.

Actions can occur when animations finish or loop, as any
AnimationListeners will be notified at those events. This makes
Animations effective timers.

The code for TextHistory was cleaned up and a new feature is available
called text demarkation. This will put black lines between rendered
messages that it displays, kind of like an outline.