Visual TextBox improvements and general functionality
Small but meaningful improvements. All are optional.
1. TextBox was given a feature to display a bar at ends where text was
cut off. Now it isn't completely ambiguous whether the box has more
information contained that is hidden in virtual space.

2. TextBox will now take the edges into account for finding the index.
Thus, if you click on the edge, or drag to the edge, then the box will
scroll in the correct direction. Arrow key selection is still functional
too.

3. MenuManager was given a method to find a Clickable from its id. This
is useful so not all later-used components have to be saved globally,
and offers the opportunity for built-in functionality like for scroll
bar buttons.

4. Panels were given a method to remove the touch responsiveness of
their child components. This is very useful if the panel is going to be
removed from shown, that its components will not clog up the touch list
held in the MenuManager.