Text Box work again
It seemed like a good time to commit now that things are stable and some
bugs were fixed. Previously there was an inconsistency between rendering
and finding the index in the text, where the former would use >= in the
calculation, and the later would use >. This has been resolved.
Another issue was that text boxes with word splitting could not be
scrolled down on because the bound detection would internally change the
mouse coordinate value. This was sufficient for boxes without word
splitting, but not for boxes with splitting, since the otherwise
consumed character takes a visible index and there is no all left
blinker position. Regardless, that issue has been resolved, and the code
for upper bound that was written in the same style, although it had yet
to cause any problems, was also altered accordingly.