Text History Patch
For better or for worse, I removed the feature where the TextHistory
would try to keep with the text that it was on before a resize occurred.
This would sometimes crash the program, and it wasn't worth the effort
anymore, especially with the variable sizes of texts.
To continue to use this feature, set the index manually with
ScrollBar.setOffset(...).