Mouse Scrolling
The main feature of this update was allowing mouse scrolling to
automatically have functionality for scrollables, like PanelPlus,
TextBox, and TextHistory. Now programmers will just have to pass the
mouse scrolling events to the MenuManager like with the other input
events. Behind the scenes, scrollable components generate a box, that
relates to screen coordinates, where they are active to scrolling. For
example, PanelPlus should only be active to scrolling when the mouse is
hovering over them. This active scrolling box matches a click boundary
quite well, insomuch that TextBox uses the same 2D array for both.
To get a little technical, the PanelPlus lies to its components about
where it is because they should render inside its independent image,
which is later drawn to the screen. After they had rendered, however,
the PanelPlus would reach inside the components and alter their click
boundaries to be correct with regards to the screen. This behavior has
been altered! Now components can effectually get their true coordinates
from their parent panel by calling getRenderOffset(...). Or, if the
components create the 2D array based on the values that they do know,
they can call handleOffsets(...) which create a 2D array for them with
corrected values.
Giving each components the freedom to get their true pixel coordinates
on the screen gives a lot of flexibility to third party developers using
this library. As long as container panels give correct offsets, children
components can be confident of their calculated true coordinates,
regardless of how the container panel implements rendering them.

A few small improvements also occurred. Since many components are
rectangularly-based, MenuComponent now provides a protected method to
them for finding their render coordinates from the given render
parameters. This method is defined as getRectRenderCoords(...). It also
supports the usage of ?location as an end point for width and height
expression strings.
TextBox was given an option to have a character mask. This could be
useful for password hiding.
An outline bug was fixed for PanelPlus.