Scrolling Pull
Most notably, I removed the "pullNegative" field in ScrollBar (since it
actually did nothing) and implemented how I had always intended pull to
be used. It is still the component creator's responsibility to call the
correct resizing functions, but there is an option: setTotalOffsets(int
total, boolean pullPositive).

Although the exact implementation of TextHistory did not allow the use
of the new method, I altered its functionality to use the same logic.

Thus, offsets will be retained, even when new elements are being added
(unless the offset is 0, in which case it is assumed the user stays up
to date).