PartitionPanel fix and new ListPanel
I fixed a silly error in PartitionPanel where it would assume that it
had no xx or yy offset when rendering. I made Draggable components a
little more user friendly by giving an error message if the programmer
fails to return a valid change array.
Lastly, I made the ListPanel class. This class is a subclass of
PanelPlus, specially designed to hold a list of elements. Since the
number of elements held is variable, the fullHeight of the panel should
change with additions or removals. The ListPanel serves this very
purpose, and since it is a subclass of PanelPlus, it is compatible with
ScrollBar and virtual space.
(I also added a couple accessor functions to GridFormatter to use in the
ListPanel class.)