Style fix-up
No new features this time. Admittedly, I got into a bad habit of using an array for dimensions. I had a de facto standard where 0 was x, 1 was y, 2 was width, and 3 was height. Unfortunately, this standard is not very readable, so I went through my code and I tried to remove it.

For the MenuComponent class, I recognized that other projects dependent on this library may have used the getRectRenderCoords(...) method. It has been deprecated and replaced with the equivalent getRenderRect(...) method.

The new standard is to return a Rectangle from awt. Conveniently, Rectangle has a constructor of (x,y,width,height), so not too many changes had to be made.