Reconfiguration
I made some changes to the structure of the library, but not really any changes to the content. I put all of the source folders in the folder "src". I did this because I intended to include my examples in the repository, and they should be kept in a separate example folder. Now they are.

I revamped the readme file.