Version 1.11
* Reconfiguration

I made some changes to the structure of the library, but not really any changes to the content. I put all of the source folders in the folder "src". I did this because I intended to include my examples in the repository, and they should be kept in a separate example folder. Now they are.

I revamped the readme file.

* Ignored in the repository

I had to remove some unintended files from the git repo.

* Documentation and Explanation

I greatly increased the available documentation for the library, particularly at a more abstracted view.

I also made a small fix in regards to radio buttons to keep them from reporting losing focus excessively.

* Fixed broken links and empty lines

I fixed a couple broken links in the Menu Components overview and I removed some unnecessary empty lines.

* Ready for MSM v1.11

Added the JavaDoc for the current project and the jar.