Introduced LineBreak
I am working on fixing a bug with TextBox and otherwise polishing up the class. I recognized an opportunity to create another class, LineBreak, that performs the functionality of finding the optimal place to break a line. This functionality was already being used by the TextBox class, as well as by TextHistory and Caption. In this update, I altered TextBox to use the LineBreak class. In future updates I will update the Caption and TextHistory classes.

Furthermore, I made some stylistic edits to TextBox as I was reviewing it. I imposed a maximum line length rule, I created temp variables to use to make the code more clear, and I decomposed a couple functions, creating the new deleteSelection(String) and getLineXOffs(String, int, int, int) methods.

In addition to cleaning up the code, these can be overriden by subclasses as necessary to promote functionality instead of hacky workarounds or rewriting sections.