Improving TextBox Rendering
I have made significant progress since the last commit. The original bug that I was trying to fix in last commit and this was that the text box was not handling the message properly if the last character was a newline and it didn't already have the row to itself because of overflow. I can proudly announce that this has been resolved (in addition to the dependent scroll bar). This came with a little restructuring of the render method and some needed simplification where possible.

However, whether I created a new bug or the bug was there before, I discovered another issue where findIndex at the bottom of the text box gives an index result that is too far down. This is most evident by the down arrow action in the TextEditBox, although it can occur in a regular TextBox.

Also in this commit, I changed the editable field in the clickable to enabled. That name makes a lot more sense, since editable applies almost only to text boxes. On the other hand, enabled is more nearly a synonym for usable, which more adequately defines the intention of the field.