Bug Shuffle
So I fixed all of the bugs that I am aware of for TextBox use when word splitting is off. Since that is the most useful mode, I consider that a success. However, I did some testing for when word splitting is off, and it crashes easily. I am still working on figuring out why and fixing the problem.

For other news, I updated the terminology in the docs for Button and PolygonalButton to reflect the change from editable -> enabled.

I also wanted to have more standardization of getter and setter names. For wordSplitting and for virtualSpace I had some strange names. It wasn't without reason since getVirtualSpace (for example) conveyed the wrong meaning. Therefore, I changed the field names where necessary. I also think that the precedence of get for getters and set for setters should override is for boolean returns, so that is why the HotkeyTextComponent had its method name changed.