Example 7
Working on Example 7, I found a couple bugs in the TextBox code that I wouldn't normally encounter, but I thought that they were worth fixing.

Additionally, I made some needed improvements to Example 7 to make it function more intuitively and keep it up to date with the changes that I have been making in the main Moulton Scalable Menus code.