LineBreak documentation and Example 7
Primarily, I added documentation for the LineBreak class and I polished up Example 7. Although I artistically chose to have the TextEditBox work the way that it did in prior updates, I have now decided that it probably is better to have the box work similarly to how other sites handle text boxes (rather than choose to be unique). It is simply an example regardless, so the any library user should not feel pressured to implement it in one way or the other. However, as it currently stands, the TextEditBox serves as a more typical example than previously.