Version 1.12
* Reconfiguration

I made some changes to the structure of the library, but not really any changes to the content. I put all of the source folders in the folder "src". I did this because I intended to include my examples in the repository, and they should be kept in a separate example folder. Now they are.

I revamped the readme file.

* Ignored in the repository

I had to remove some unintended files from the git repo.

* Documentation and Explanation

I greatly increased the available documentation for the library, particularly at a more abstracted view.

I also made a small fix in regards to radio buttons to keep them from reporting losing focus excessively.

* Fixed broken links and empty lines

I fixed a couple broken links in the Menu Components overview and I removed some unnecessary empty lines.

* Ready for MSM v1.11

Added the JavaDoc for the current project and the jar.

* Removed redundant jar

I removed the redundant .jar because I already have it downloaded on GitHub through the release.

* Introduced LineBreak

I am working on fixing a bug with TextBox and otherwise polishing up the class. I recognized an opportunity to create another class, LineBreak, that performs the functionality of finding the optimal place to break a line. This functionality was already being used by the TextBox class, as well as by TextHistory and Caption. In this update, I altered TextBox to use the LineBreak class. In future updates I will update the Caption and TextHistory classes.

Furthermore, I made some stylistic edits to TextBox as I was reviewing it. I imposed a maximum line length rule, I created temp variables to use to make the code more clear, and I decomposed a couple functions, creating the new deleteSelection(String) and getLineXOffs(String, int, int, int) methods.

In addition to cleaning up the code, these can be overriden by subclasses as necessary to promote functionality instead of hacky workarounds or rewriting sections.

* Improving TextBox Rendering

I have made significant progress since the last commit. The original bug that I was trying to fix in last commit and this was that the text box was not handling the message properly if the last character was a newline and it didn't already have the row to itself because of overflow. I can proudly announce that this has been resolved (in addition to the dependent scroll bar). This came with a little restructuring of the render method and some needed simplification where possible.

However, whether I created a new bug or the bug was there before, I discovered another issue where findIndex at the bottom of the text box gives an index result that is too far down. This is most evident by the down arrow action in the TextEditBox, although it can occur in a regular TextBox.

Also in this commit, I changed the editable field in the clickable to enabled. That name makes a lot more sense, since editable applies almost only to text boxes. On the other hand, enabled is more nearly a synonym for usable, which more adequately defines the intention of the field.

* Bug Shuffle

So I fixed all of the bugs that I am aware of for TextBox use when word splitting is off. Since that is the most useful mode, I consider that a success. However, I did some testing for when word splitting is off, and it crashes easily. I am still working on figuring out why and fixing the problem.

For other news, I updated the terminology in the docs for Button and PolygonalButton to reflect the change from editable -> enabled.

I also wanted to have more standardization of getter and setter names. For wordSplitting and for virtualSpace I had some strange names. It wasn't without reason since getVirtualSpace (for example) conveyed the wrong meaning. Therefore, I changed the field names where necessary. I also think that the precedence of get for getters and set for setters should override is for boolean returns, so that is why the HotkeyTextComponent had its method name changed.

* Example 7

Working on Example 7, I found a couple bugs in the TextBox code that I wouldn't normally encounter, but I thought that they were worth fixing.

Additionally, I made some needed improvements to Example 7 to make it function more intuitively and keep it up to date with the changes that I have been making in the main Moulton Scalable Menus code.

* Polishing Up Example 7

I made some small changes to example 7 to improve its overall functionality and aesthetic.

* LineBreak documentation and Example 7

Primarily, I added documentation for the LineBreak class and I polished up Example 7. Although I artistically chose to have the TextEditBox work the way that it did in prior updates, I have now decided that it probably is better to have the box work similarly to how other sites handle text boxes (rather than choose to be unique). It is simply an example regardless, so the any library user should not feel pressured to implement it in one way or the other. However, as it currently stands, the TextEditBox serves as a more typical example than previously.

* Polishing for the next release version

I did some slight polishing for the next release version, and though there were a lot of altered files, a great deal of it is documentation file changes.