Caught a silly bug
I don't know exactly how I missed this issue before. The logic behind it was very unclear, I assume that it was left behind as an artifact from when I transitioned the functioning of the TextEditBox to be more in line with common usage.
The text itself was not wrong, just misplaced.