Silly fixes and image padding for image buttons
I fixed a few silly issues like a typo in the readme file and the unnecessary string addition in the MenuManager. I also fixed an issue with TextBox that hopefully has not been present for too long. It was about inserting characters in the middle of a message. It would fail after a paste or set message because I naively assumed that said two operations would always want to put the index at the end of the set message.
But the big advancement of this update is the padding for the ImageButton. I didn't make this earlier since I figured the changes could be done with the reference image file, but I figured out of convenience I might as well add it.