Mouse Reactivity
I finally decided to add mouse cursor reactivity. This is something that I have been thinking about for years. In the beginning, it didn't make sense to add it since the touch sensitivity was contrary to my minimalist approach. However, touch sensitivity is fully integrated and here to stay, so I decided that I ought to provide the opportunity for cursor reactivity.
As said, it is still optional. Cursor reactivity only makes sense in the context of touch sensitivity, so if components are not touch sensitive, they will not affect the cursor. Furthermore, touch sensitive components can opt out by setting their cursorType as the default, which will thus have no effect.
Also a small note. I already made the EventAction class work properly, but it appears that I forgot to include it with the repository in previous updates. Regardless, it is here now.

The mouse cursor reactivity changes required me to change my examples slightly.