Cursors, Statics, and Panels
There were some important changes. I will address them in reverse order from the title of this change.

First, there was a small change made to panels. It turns out that even though I had declared lastWidth for Panel, I forgot to set it in render. This has been fixed. For PanelPlus (which may soon change names to VirtualPanel to be more specific), it turns out that rendering did not utilize the corrective measures for its coordinates, meaning that prior to this update, a PanelPlus would have an incorrect location if nested. I have patched this, but not tested it yet. Finalization will take place in a future update before the next release.

Second, the static field of deselectOnRelease has been removed from Clickable. It was actually unused, and static inheritance in Java does not support polymorphism, so it would never be useful. If a subclass wants to change deselect on release behavior, they can override the method.

Going along with this is a change to cursor implementation. I was thinking of using it about the same as deselect on release, so I decided to change both. There is no longer a protected cursorType field in Clickable. If a subclass wants to return a different cursorType value, they can override the getter method. This comes with a few negative consequences, but the uses should be rare. For example, if the cursor type getter in a parent class A, did work, and if the subclass of A, B, wanted to return a different cursorType, then the getter would be overridden and the work would not be done. Of course, B could call super and disregard the result, but enforcing a super call for all children is a brittle system. It appears that it is currently unavoidable.

Lastly, the way that the MenuManager handles cursor types has been improved upon. If a touch-sensitive type does not want to modify the cursor, it can simply return Cursor.DEFAULT_CURSOR, and the manager will disregard it. Otherwise, if there is a conflict, the most recently added to the touch list will take precedence (as the updated documentation specifies).