Version 1.13
* Expression Max and Min

I made some small additions to ExpressionSolver where it can now solve expressions using the max and/or min functions. I also fixed a small bug with ExpressionSolver that sometimes occured when several sets of parentheses were present.

* Caught a silly bug

Caught a silly insertion glitch with the TextBox.

* Documentation changes

Expanded the documentation and fixed typos.

* Image padding for image buttons

But the big advancement of this update is the padding for the ImageButton. I didn't make this earlier since I figured the changes could be done with the reference image file, but I figured out of convenience I might as well add it.

* Mouse Reactivity

I finally decided to add mouse cursor reactivity. This is something that I have been thinking about for years. In the beginning, it didn't make sense to add it since the touch sensitivity was contrary to my minimalist approach. However, touch sensitivity is fully integrated and here to stay, so I decided that I ought to provide the opportunity for cursor reactivity.
As said, it is still optional. Cursor reactivity only makes sense in the context of touch sensitivity, so if components are not touch sensitive, they will not affect the cursor. Furthermore, touch sensitive components can opt out by setting their cursorType as the default, which will thus have no effect.
Also a small note. I already made the EventAction class work properly, but it appears that I forgot to include it with the repository in previous updates. Regardless, it is here now.

The mouse cursor reactivity changes required me to change my examples slightly.

* Cursors, Statics, and Panels

There were some important changes. I will address them in reverse order from the title of this change.

First, there was a small change made to panels. It turns out that even though I had declared lastWidth for Panel, I forgot to set it in render. This has been fixed. For PanelPlus (which may soon change names to VirtualPanel to be more specific), it turns out that rendering did not utilize the corrective measures for its coordinates, meaning that prior to this update, a PanelPlus would have an incorrect location if nested. I have patched this, but not tested it yet. Finalization will take place in a future update before the next release.

Second, the static field of deselectOnRelease has been removed from Clickable. It was actually unused, and static inheritance in Java does not support polymorphism, so it would never be useful. If a subclass wants to change deselect on release behavior, they can override the method.

Going along with this is a change to cursor implementation. I was thinking of using it about the same as deselect on release, so I decided to change both. There is no longer a protected cursorType field in Clickable. If a subclass wants to return a different cursorType value, they can override the getter method. This comes with a few negative consequences, but the uses should be rare. For example, if the cursor type getter in a parent class A, did work, and if the subclass of A, B, wanted to return a different cursorType, then the getter would be overridden and the work would not be done. Of course, B could call super and disregard the result, but enforcing a super call for all children is a brittle system. It appears that it is currently unavoidable.

Lastly, the way that the MenuManager handles cursor types has been improved upon. If a touch-sensitive type does not want to modify the cursor, it can simply return Cursor.DEFAULT_CURSOR, and the manager will disregard it. Otherwise, if there is a conflict, the most recently added to the touch list will take precedence (as the updated documentation specifies).

* Renames

I refactored a couple renames (getClicked -> setClicked, PanelPlus -> VirtualPanel) to be more descriptive and match standards better.

* FormButton

A button subclass specially made to be integrated into forms. The form button can recognize enter characters when selected to act as if it was clicked.

* Remove JavaDocs from Repo

I have decided to put the include the Javadocs in the release by means of a downloadable zip file. It didn't make sense to me to include the formal docs in with the rest of the repo files since they change often and bloat the source download.