Expression standardization and improvements
I made some improvements to expression solving. First of all, I fixed a bug with parentheses in the ExpressionSolver class.
Second, I was able to add 4 new variables to rectangular components. They are CENTERX, CENTERY, WIDTH, and HEIGHT. Instead of being relative to the canvas space (like the lowercase variables), the uppercase variables are relative to the component dimensions. This allows for simpler calculations.
Third, I performed some expression standardization. A lot of components did their own logic for the expressions, which was passed down from very old code. However, there were many redundancies in this code. Therefore, the code was moved to getRenderRect in MenuComponent for all to share.