Example Update
I updated a couple of the examples and I uploaded the newer Example 8. Furthermore, I updated the documentation md files to specify the new variables and explain the new example.

Lastly, I had to update the ExpressionSolver again. It should be more robust now after another fix to parentheses and a fix to account for multiple negations in a string beginning with a leading negative.