TextFormat changes
Made some changes to TextFormat to allow for more complicated formats. Specifically, the validity of a character may depend on the message prior, which requires message to be in a known state before each call to IsValidChar and for each character to be processed sequentially.

Also made some non-breaking API changes for set methods where this is returned. This should allow for easier and/or cleaner construction.